경로 재수정
