+++ server/src/entity/Entity.ts
@@ -1,0 +1,12 @@
+import { BaseEntity, CreateDateColumn, PrimaryGeneratedColumn, UpdateDateColumn} from "typeorm"
+
+export default abstract class Entity extends BaseEntity {
+    @PrimaryGeneratedColumn()
+    id: number;
+
+    @CreateDateColumn()
+    creadtedAt: Date;
+
+    @UpdateDateColumn()
+    updatedAt: Date;
+}
