+++ server/src/entity/MainBanner.ts
