+++ server/src/entity/Product.ts
