+++ server/src/entity/Style.ts
